feat(finding): add Confidence named type replacing bare float64
Finding.Confidence and Correlation.Confidence changed from float64 to
the new Confidence type, a named float64 with validation.

New type: Confidence (confidence.go)
- ConfidenceNone (0.0), ConfidenceLow (0.25), ConfidenceMedium (0.5),
  ConfidenceHigh (0.75), ConfidenceFull (1.0)
- IsValid() — checks [0.0, 1.0] range
- Clamp() — returns value clamped to [0.0, 1.0]
- String() — formatted representation

Changes:
- Finding.Confidence: float64 → Confidence
- Correlation.Confidence: float64 → Confidence
- NewFinding: keeps float64 param for convenience, converts internally
- Builder.WithConfidence: keeps float64 param, converts internally
- NormalizedConfidence: returns Confidence instead of float64
- Validate: uses Confidence.String() in error messages
- Equal: uses direct == comparison (exact, no epsilon needed for
  values that went through Clamp)
- SARIF export: explicit float64() cast for Rank field
- SARIF import: Confidence() cast for rank and property conversion
- Removed clampConfidence() free function (replaced by Clamp() method)
- Updated TestClampConfidence to test Confidence.Clamp() method

Compiler-enforced benefit: Finding{Confidence: 1.5} still compiles
(untyped constant), but passing a float64 variable now requires
explicit Confidence(x) conversion — making accidental out-of-range
values visible at the call site.

Build: GREEN | Tests: 502 PASS | Vet: CLEAN

+++ sarif_export.go
@@ -130,7 +130,7 @@ func findingToSARIF(f Finding) SarifResult {
 				},
 			},
 		}},
-		Rank: f.NormalizedConfidence() * sarifConfidenceScale,
+		Rank: float64(f.NormalizedConfidence()) * sarifConfidenceScale,
 	}
 
 	if f.Range != nil && f.Range.HasEnd() {
